feat(notif): mise en place de la notification lorsque le taux d exposition est trop eleve
