+++ android/app/src/main/AndroidManifest.xml
@@ -18,6 +18,7 @@
             android:enabled="true">
             <intent-filter>
                 <action android:name="android.intent.action.BOOT_COMPLETED"/>
+                <action android:name="android.intent.action.MY_PACKAGE_REPLACED"/>
             </intent-filter>
         </receiver>
         <meta-data android:name="com.google.android.geo.API_KEY"
