+++ pubspec.yaml
@@ -46,6 +46,7 @@ dependencies:
   shared_preferences: ^0.5.10
   geocoding: ^1.0.5
   geolocator: ^5.0.0
+  flutter_local_notifications: ^1.4.4+2
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
